Backend/requirements.txt
fastapi>=0.112.0,<1.0.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
psycopg2-binary>=2.0.0,<3.0.0
typing>=3.0.0,<4.0.0
python-multipart>=0.0.0,<1.0.0
passlib[bcrypt]==1.7.4